backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
openai==1.50.0
sentence-transformers==3.0.0
umap-learn==0.5.6
numpy==1.26.4
scikit-learn==1.5.0
python-dotenv==1.0.0
